+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -21,6 +21,7 @@
 #include "Rule15p5Check.h"
 #include "Rule15p6Check.h"
 #include "Rule15p7Check.h"
+#include "Rule16p2Check.h"
 #include "Rule16p4Check.h"
 #include "Rule16p5Check.h"
 #include "Rule16p7Check.h"
@@ -66,6 +67,8 @@ public:
         "misrac-rule-15p6");
     CheckFactories.registerCheck<Rule15p7Check>(
         "misrac-rule-15p7");
+    CheckFactories.registerCheck<Rule16p2Check>(
+        "misrac-rule-16p2");
     CheckFactories.registerCheck<Rule16p4Check>(
         "misrac-rule-16p4");
     CheckFactories.registerCheck<Rule16p5Check>(